+++ rastervision_core/rastervision/core/data/raster_source/__init__.py
@@ -6,3 +6,5 @@ from rastervision.core.data.raster_source.rasterio_source import *
 from rastervision.core.data.raster_source.rasterio_source_config import *
 from rastervision.core.data.raster_source.rasterized_source import *
 from rastervision.core.data.raster_source.rasterized_source_config import *
+from rastervision.core.data.raster_source.multi_raster_source import *
+from rastervision.core.data.raster_source.multi_raster_source_config import *