+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Features
 * Added support for multiband images `#972 <https://github.com/azavea/raster-vision/pull/972>`_
 * Add support for vector output to predict command `#980 <https://github.com/azavea/raster-vision/pull/980>`_
 * Add support for weighted loss for classification and semantic segmentation `#977 <https://github.com/azavea/raster-vision/pull/977>`_
+* Add multi raster source `#978 <https://github.com/azavea/raster-vision/pull/982>`_
 
 Raster Vision 0.12
 -------------------